+++ Carbon.Cryptography/Carbon.Cryptography.psm1
@@ -25,7 +25,8 @@ $moduleBinRoot = Join-Path -Path $moduleRoot -ChildPath 'bin'
 $moduleBinRoot | Out-Null # To make the PSScriptAnalyzer squiggle go away.
 $privateModulesRoot = Join-Path -Path $moduleRoot -ChildPath 'Modules'
 
-Import-Module -Name (Join-Path -Path $privateModulesRoot -ChildPath 'Carbon.Core')
+Import-Module -Name (Join-Path -Path $privateModulesRoot -ChildPath 'Carbon.Core') `
+              -Function @('Invoke-CPowerShell', 'Test-COperatingSystem')
 
 # Store each of your module's functions in its own file in the Functions 
 # directory. On the build server, your module's functions will be appended to