+++ Carbon.Cryptography/bin/Protect-String.ps1
@@ -15,16 +15,14 @@
 # limitations under the License.
 param(
     [Parameter(Mandatory)]
-    [String]$ProtectedString
+    [String] $ProtectedString
 )
 
 #Requires -Version 5.1
 Set-StrictMode -Version 'Latest'
 
-# Keep cause this script is used by PowerShell 2.
-Add-Type -AssemblyName 'System.Security'
-
-Import-Module -Name (Join-Path -Path $PSScriptRoot -ChildPath '..\Carbon.Cryptography.psd1' -Resolve)
+Import-Module -Name (Join-Path -Path $PSScriptRoot -ChildPath '..\Carbon.Cryptography.psd1' -Resolve) `
+              -Function @('Unprotect-String', 'Protect-String')
 
 $string = Unprotect-CString -ProtectedString $ProtectedString
 Protect-CString -String $string -ForUser