+++ Tests/Initialize-Test.ps1
@@ -34,7 +34,9 @@ if( (Get-Module -Name 'Carbon') )
 }
 
 Import-Module -Name 'Microsoft.PowerShell.Security'
-Import-Module -Name (Join-Path -Path $PSScriptRoot -ChildPath '..\PSModules\Carbon.Core' -Resolve)
+Import-Module -Name (Join-Path -Path $PSScriptRoot -ChildPath '..\PSModules\Carbon.Core' -Resolve) `
+              -Prefix 'T' `
+              -Force
 
 try
 {